+++ regionmask/core/mask.py
@@ -248,13 +248,14 @@ def _mask_3D(
 
     # if no regions are found return a 0 x lat x lon mask
     if len(numbers) == 0:
-        mask = mask.expand_dims("region", axis=0).sel(region=slice(0, 0))
+        mask_3D = mask.expand_dims("region", axis=0).sel(region=slice(0, 0))
+        mask_3D = mask_3D.assign_coords(region=("region", numbers))
         msg = (
             "No gridpoint belongs to any region. Returning an empty mask"
-            " with shape {}".format(mask.shape)
+            f" with shape {mask.shape}"
         )
         warnings.warn(msg, UserWarning, stacklevel=3)
-        return mask
+        return mask_3D
 
     mask_3D = list()
     for num in numbers: