+++ docs/whats_new.rst
@@ -73,6 +73,8 @@ Internal Changes
   download URL and catch all ``URLError``, not only timeouts (:pull:`289`).
 - Rewrote the function to create the mask `DataArray` (:issue:`168`, :pull:`294`).
 - Follow up to :pull:`294` - fix wrong dimension order for certain conditions (:issue:`295`).
+- Refactor `test_mask` - make use of ``xr.testing.assert_equal`` and simplify some
+ elements (:pull:`297`).
 
 v0.8.0 (08.09.2021)
 -------------------